refactor(server-core): split the theme module into contract and serving
`adapters/http/ui/theme/` had grown to seven files mixing two concerns:
what a theme IS (manifest type, validator, token grammar, asset kinds,
head builder) and how it is SERVED (filesystem provider, http handler,
response headers).

The first half now lives in `theme/contract/` and imports nothing from
node or routup — only validup, zod and @authup/errors. That makes it
portable: a browser-side theme editor or a CLI validator can consume it
verbatim, and plan 085's @authup/theme-kit extraction becomes a directory
move rather than a rewrite. `themeAssetExtension` replaces `node:path`'s
extname, which was the one thing breaking the rule.

Consequently the constants split too. contract/ owns the on-disk layout,
the manifest version, the token grammar and the asset-kind table (the
extension allowlist is derived from the content-type map, so they cannot
drift); serving keeps the revalidate interval and the asset CSP.
`applyTheme` moved out of the head builder into `theme/apply.ts`, since it
takes a filesystem-backed provider and splices a served document, leaving
`contract/head.ts` pure.

`ui/contract.ts` also became `ui/console-packages/`, matching the
module.ts + types.ts + index.ts shape of its auth-console and
account-console siblings and removing the name collision with the new
theme contract folder.

No behaviour change: pure moves plus import rewiring.
