+++ apps/server-core/src/adapters/http/ui/console-packages/index.ts
@@ -5,10 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './contract/index.ts';
-
-export * from './apply.ts';
-export * from './assets.ts';
-export * from './constants.ts';
 export * from './module.ts';
 export * from './types.ts';
