+++ apps/server-core/src/adapters/http/ui/index.ts
@@ -7,7 +7,6 @@
 
 export * from './account-console/index.ts';
 export * from './auth-console/index.ts';
-export * from './contract.ts';
+export * from './console-packages/index.ts';
 export * from './shared/index.ts';
 export * from './theme/index.ts';
-export * from './types.ts';
