+++ apps/server-core/src/adapters/http/ui/theme/assets.ts
@@ -10,11 +10,8 @@ import path from 'node:path';
 import { Readable } from 'node:stream';
 import type { Handler } from 'routup';
 import { HeaderName, defineCoreHandler, sendFile } from 'routup';
-import {
-    THEME_ASSET_CONTENT_TYPES,
-    THEME_ASSET_CSP,
-    THEME_ASSET_EXTENSIONS,
-} from './constants.ts';
+import { THEME_ASSET_CSP } from './constants.ts';
+import { THEME_ASSET_CONTENT_TYPES, THEME_ASSET_EXTENSIONS } from './contract/index.ts';
 import type { IThemeProvider } from './types.ts';
 
 const REQUEST_PATH_PATTERN = /^[a-z0-9][a-z0-9._/-]*$/i;
