+++ apps/server-core/src/adapters/http/ui/theme/contract/head.ts
@@ -5,16 +5,15 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { injectHeadContent, stampDocumentTitle } from '../shared/index.ts';
 import {
     THEME_ASSETS_DIRECTORY_NAME,
     THEME_ASSET_MOUNT_PATH,
     THEME_CSS_LAYER_NAME,
     THEME_LOGO_TOKENS,
 } from './constants.ts';
-import type { IThemeProvider, ThemeManifest } from './types.ts';
+import type { ThemeManifest } from './types.ts';
 
-function escapeHtml(value: string) : string {
+export function escapeHtml(value: string) : string {
     return value
         .replace(/&/g, '&amp;')
         .replace(/</g, '&lt;')
@@ -105,35 +104,4 @@ export function buildThemeHead(
     }
 
     return parts.join('');
-}
-
-/**
- * Apply the operator theme to a rendered console shell. A no-op when
- * theming is off or the directory carries no manifest, so both console
- * serve paths can call it unconditionally.
- */
-export async function applyTheme(
-    html: string,
-    provider: IThemeProvider | undefined,
-    basePath: string,
-) : Promise<string> {
-    if (!provider) {
-        return html;
-    }
-
-    let body = html;
-
-    // A directory may carry only a fragment, or only a stylesheet, so the
-    // manifest is not a precondition for applying the theme.
-    const manifest = await provider.getManifest();
-    if (manifest?.title) {
-        body = stampDocumentTitle(body, escapeHtml(manifest.title));
-    }
-
-    const head = await provider.getHead(basePath);
-    if (head) {
-        body = injectHeadContent(body, head);
-    }
-
-    return body;
 }
