+++ apps/server-core/src/adapters/http/ui/theme/contract/index.ts
@@ -5,10 +5,8 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './contract/index.ts';
-
-export * from './apply.ts';
-export * from './assets.ts';
 export * from './constants.ts';
-export * from './module.ts';
+export * from './head.ts';
+export * from './manifest.ts';
 export * from './types.ts';
+export * from './utils.ts';
