+++ apps/server-core/src/adapters/http/ui/theme/contract/types.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2026.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+/**
+ * The operator-authored `theme.json`, after validation.
+ *
+ * `favicon` / `logo` / `stylesheet` are theme-relative paths inside
+ * `assets/`; they are emitted as `<basePath>/theme/<name>` hrefs.
+ */
+export type ThemeManifest = {
+    version: number,
+    title?: string,
+    favicon?: string,
+
+    /**
+     * Replaces the built-in logo mark on both consoles. Maps onto the
+     * `--authup-{auth,account}-logo-*` token pairs.
+     */
+    logo?: string,
+
+    stylesheet?: string,
+    tokens?: Record<string, string>,
+    tokensDark?: Record<string, string>,
+};
